// src/definitions/interfaces/interface.ts
export type SearchParams = {
  [x: string]: any;
  pageSize?: number;
  pageIndex?: number;
  keyword?: string;
  activityTypeId?: string;
  status?: string | string[];
  state?: string;
  sortType?: 'ASC' | 'DESC';
  sortBy?: string;
  type?: string;
  companyIds?: string[];
  locationCodes?: string[];
  industryCodes?: string[];
  download?: 'xlsx';
  excelHeaders?: Record<string, string>;
  platformRole?: string;
  activityRoles?: string;
  location?: string;
  more?: {
    jobTitle?: string;
    relationship?: string;
    skillCategory?: string;
    employmentType?: string;
    name?: string;
    description?: string;
    serviceType?: string;
    incorporationYear?: string;
    serviceOffering?: string;
    platformRole?: string;
    classification?: string;
    applicationId?: string;
    createdAt?: string;
    enrolledTime?: string;
    deadlineSubmit?: string;
  };
  developmentStage?: string;
  page?: number;
  limit?: number;
  s?: string;
  columns?: any;
  companyType?: any;
  inActivityId?: string;
  messageTime?: string;
  conversationType?: string;
  roundIds?: any;
  paticipants?: number;
};